refactor: move internal/control/ to internal/muxcore/control/ (T008)
