+++ cmd/mcp-mux/daemon.go
@@ -10,7 +10,7 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/thebtf/mcp-mux/internal/control"
+	"github.com/thebtf/mcp-mux/internal/muxcore/control"
 	"github.com/thebtf/mcp-mux/internal/daemon"
 	"github.com/thebtf/mcp-mux/internal/muxcore/ipc"
 	"github.com/thebtf/mcp-mux/internal/muxcore/serverid"
