+++ internal/daemon/snapshot_test.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/thebtf/mcp-mux/internal/muxcore/classify"
-	"github.com/thebtf/mcp-mux/internal/control"
+	"github.com/thebtf/mcp-mux/internal/muxcore/control"
 	"github.com/thebtf/mcp-mux/internal/mux"
 )
