+++ internal/mcpserver/server_test.go
@@ -12,7 +12,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/thebtf/mcp-mux/internal/control"
+	"github.com/thebtf/mcp-mux/internal/muxcore/control"
 )
 
 // newTestServer creates a Server wired to an io.Pipe pair and returns
