+++ testdata/smoke_isolated.go
@@ -33,7 +33,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/thebtf/mcp-mux/internal/control"
+	"github.com/thebtf/mcp-mux/internal/muxcore/control"
 	"github.com/thebtf/mcp-mux/internal/muxcore/serverid"
 )
